refactor(tasks): separate flow record transformations
